Civic-Ai/requirements.txt
fastapi
uvicorn
python-multipart
motor
python-dotenv
pydantic
email-validator
requests
geopy
fastapi-mail
passlib
bcrypt==4.0.1
timm
albumentations
opencv-python
pillow
numpy
scikit-learn
google-generativeai
sentence-transformers
dnspython